[Xbox]: merged Linuxport revisions 16964,16966,16967,16973,16975,16978-16980,16982,16983,16985-16992,16994,16998 17003-17005,17008,17009,17012,17013,17016,17018-17021,17024-17030,17032-17035 17037-17040,17042-17045,17055,17057-17061,17063-17068
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@17069 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/utils/MusicInfoScraper.cpp
@@ -172,8 +172,8 @@ void CMusicInfoScraper::FindAlbuminfo()
         float scale=1;
         const char* newscale = relevance->Attribute("scale");
         if (newscale)
-          scale = atof(newscale); 
-        newAlbum.SetRelevance(atof(relevance->FirstChild()->Value())/scale);
+          scale = (float)atof(newscale);
+        newAlbum.SetRelevance((float)atof(relevance->FirstChild()->Value())/scale);
       }
       m_vecAlbums.push_back(newAlbum);
     }
